Change the type of `id`
At first, we define 3 main types with their `id` field to be optional in
order to satisfy insertion operation. However, it is wrong because the
database schemas are always `NOT NULL`, which makes it more complicated
to code. Now, we change the type of `id` with no `undefined`, and `Omit`
the `id` field when making insertion.